changed id in entities from int to Long

+++ JSF-Forum-ejb/src/java/jsf/forum/controller/CategoryController.java
@@ -27,7 +27,7 @@ public class CategoryController {
         em.persist(category);
     }
 
-    public Category getCategoryById(int id) {
+    public Category getCategoryById(Long id) {
         Category category = em.find(Category.class, id);
         return category;
     }
